reorganized scene file structures
