+++ src/App.js
@@ -1,6 +1,6 @@
-import React, { Fragment, Component } from 'react'
+import React, { Fragment, Component } from 'react';
 // router
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 // bootstrap
 // import Container from 'react-bootstrap/Container';
@@ -8,22 +8,22 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 // import Col from 'react-bootstrap/Col';
 // import Image from 'react-bootstrap/Image';
 // pages
-import Landing from './components/scenes/Landing'
-import About from './components/scenes/About'
-import Meetups from './components/scenes/Meetups'
-import Jobs from './components/scenes/Jobs'
-import Resources from './components/scenes/Resources'
-import Testimonials from './components/scenes/Testimonials'
-import NonProfit from './components/scenes/NonProfit'
+import Landing from './components/scenes/Landing/index';
+import About from './components/scenes/About/index';
+import Meetups from './components/scenes/Meetups/index';
+import Jobs from './components/scenes/Jobs/index';
+import Resources from './components/scenes/Resources/index';
+import Testimonials from './components/scenes/Testimonials/index';
+import NonProfit from './components/scenes/NonProfit/index';
 
 // components
-import NavBar from './components/NavBar'
-import Volunteer from './components/scenes/Volunteer'
-import withTracker from './tracker/withTracker'
+import NavBar from './components/NavBar';
+import Volunteer from './components/scenes/Volunteer';
+import withTracker from './tracker/withTracker';
 
 // css
-import 'bootstrap/dist/css/bootstrap.min.css'
-import './App.scss'
+import 'bootstrap/dist/css/bootstrap.min.css';
+import './App.scss';
 
 // images
 class App extends Component {
@@ -45,8 +45,8 @@ class App extends Component {
 					</Switch>
 				</Router>
 			</Fragment>
-		)
+		);
 	}
 }
 
-export default App
+export default App;
