+++ src/components/scenes/About/style.scss
@@ -1,3 +1,5 @@
+@import '../../../App.scss';
+
 .single {
 	margin-bottom: 32px;
 }
