+++ src/components/scenes/Blog/Blog.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './style.scss';
 
 const Blog = () => {
 	return (
