+++ src/components/scenes/Blog/style.scss
@@ -1,0 +1,1 @@
+@import '../../../App.scss';
