+++ src/components/scenes/Jobs/index.js
@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react';
+import './style.scss';
 
-import Footer from '../Footer';
+import Footer from '../../Footer';
 const Jobs = () => {
 	return (
 		<Fragment>
