+++ src/components/scenes/Jobs/style.scss
@@ -1,0 +1,1 @@
+@import '../../../App.scss';
