+++ src/components/scenes/Meetups/index.js
@@ -1,12 +1,12 @@
 import React, { Fragment } from 'react';
-import Header from '../Header';
-import Footer from '../Footer';
-import MeetupMap from '../MeetupMap';
+import Header from '../../Header';
+import Footer from '../../Footer';
+import MeetupMap from '../../MeetupMap';
 import Container from 'react-bootstrap/Container';
 // import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import Meetup from '../../img/icon/Meetup.svg';
+import Meetup from './Meetup.svg';
 import './style.scss';
 
 const Meetups = () => {
