+++ src/components/scenes/Meetups/style.scss
@@ -1,0 +1,13 @@
+// TODO: Move to my own component
+.meetups .meetup-logo {
+	width: 100px;
+}
+
+.meetups .meetup-link {
+	margin: 15px 0;
+}
+
+.meetups .meetup-location-title {
+	margin-left: 25px;
+	font-size: 2em;
+}
