+++ src/components/scenes/NonProfit/index.js
@@ -1,10 +1,11 @@
 import React, { Fragment } from 'react';
-import Header from '../Header';
-import Footer from '../Footer';
+import Header from '../../Header';
+import Footer from '../../Footer';
 // import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Container from 'react-bootstrap/Container';
+import './style.scss';
 
 const NonProfit = () => {
 	return (
