+++ src/components/scenes/Resources/index.js
@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react';
 
-import Footer from '../Footer';
+import Footer from '../../Footer';
+import './style.scss';
 
 const Resources = () => {
 	return (
