+++ src/components/scenes/Testimonials/index.js
@@ -1,15 +1,15 @@
-import React, { Fragment } from 'react'
-import Header from '../Header'
-import Footer from '../Footer'
+import React, { Fragment } from 'react';
+import Header from '../../Header';
+import Footer from '../../Footer';
 // import Col from 'react-bootstrap/Col';
-import Row from 'react-bootstrap/Row'
-import Col from 'react-bootstrap/Col'
-import Container from 'react-bootstrap/Container'
-import Media from 'react-bootstrap/Media'
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import Container from 'react-bootstrap/Container';
+import Media from 'react-bootstrap/Media';
 
-import TaiMaiProfile from '../../img/testimonials/tye-mai/tye-mai.jpeg'
-import testimonoials from './testimonials.scss'
-import linkedInLogo from '../../img/icon/linkedin-icon.svg'
+import TaiMaiProfile from './img/tye-mai.jpeg';
+import './style.scss';
+import linkedInLogo from './img/linkedin-icon.svg';
 
 const Testimonials = () => {
 	return (
@@ -133,7 +133,7 @@ const Testimonials = () => {
 
 			<Footer />
 		</Fragment>
-	)
-}
+	);
+};
 
-export default Testimonials
+export default Testimonials;
